requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
certifi==2019.11.28       # via python-telegram-bot
cffi==1.14.0              # via cryptography
cryptography==2.8         # via python-telegram-bot
decorator==4.4.1          # via python-telegram-bot
future==0.18.2            # via python-telegram-bot
pycparser==2.19           # via cffi
python-telegram-bot==12.4.1
six==1.14.0               # via cryptography
tornado==6.0.3            # via python-telegram-bot
